refactor(auth): remove duplicate AuthenticatedUser from module-infra
All references redirected to module-core version.
Two identical copies existed differing only in package name.
